feat(core): payment plans skeleton (#4715)
